[workspace/api-keys] - feature: implement API key creation and management components
- Add a new sheet component to display information for a newly created API key
- Create a list component to display and manage existing API keys
- Implement a dialog component to facilitate creation of new API keys
- Extract utility functions for group name prettification into a separate utils module
- Refactor API keys page to integrate new components for better user experience
